+++ frontend/src/assets/icons/astro/ArrowRight.astro
