+++ frontend/src/assets/icons/astro/Play.astro
