+++ frontend/src/assets/icons/react/Plus.tsx
