+++ frontend/src/assets/icons/react/SearchIcon.tsx
