+++ frontend/src/components/shared/astro-components/announcement-banner.astro
