+++ frontend/src/components/shared/astro-components/hero-bg-vector.astro
