+++ frontend/src/layouts/page-layout.astro
@@ -1,7 +1,7 @@
 ---
 import "@fontsource/geist-sans";
 import "@/styles/global.css";
-import AnnouncementBanner from "@/components/shared/announcement-banner.astro";
+import AnnouncementBanner from "@/components/shared/astro-components/announcement-banner.astro";
 export interface Props {
   title?: string;
   description?: string;
