+++ frontend/src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
-import Hero from "@/components/home/hero.astro";
-import AboutOSM from "@/components/home/about-osm.astro";
-import Layout from "@/layouts/Layout.astro";
+import Hero from "@/components/home/astro-components/hero.astro";
+import AboutOSM from "@/components/home/astro-components/about-osm.astro";
+import Layout from "@/layouts/page-layout.astro";
 import { APP_CONFIG } from "@/config/Content";
 ---
