+++ frontend/src/pages/resources.astro
@@ -1,8 +1,8 @@
 ---
-import Hero from "@/components/resources/hero.astro";
-import Layout from "@/layouts/Layout.astro";
+import Hero from "@/components/resources/astro-components/hero.astro";
+import Layout from "@/layouts/page-layout.astro";
 import { APP_CONFIG } from "@/config/Content";
-import ResourceSearch from "@/components/react-components/resources/ResourceSearch";
+import ResourceSearch from "@/components/resources/react-components/ResourceSearch";
 ---
 
 <Layout
